// src/app/core/api/model/logon-response-dto.ts
/**
 * Zidium Front
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CurrentUserDto } from './current-user-dto';


export interface LogonResponseDto { 
    token: string;
    refreshToken: string;
    user: CurrentUserDto;
}